requirements.txt
requests==2.27.1
urllib3==1.26.8
python-dotenv==0.9.1
